+++ clases/Incidencias.cs
@@ -8,12 +8,12 @@ namespace clases
 {
     public class Incidencias
     {
-        public Int64 Id                  { get; set; }
+        public int Id                  { get; set; }
         public int IdCliente           { get; set; }
         public Estados  Estado            { get; set; }
         public Prioridades  Prioridad     { get; set; }
         public TiposIncidencias   TipoIncidencia    { get; set; }
-        public Int64 IdUsuario           { get; set; }
+        public int IdUsuario           { get; set; }
         public string   Problemática        { get; set; }
         public string   Comentarios         { get; set; }
     }
