+++ clases/Usuarios.cs
@@ -8,11 +8,11 @@ namespace clases
 {
     public class Usuarios
     {
-        public Int64      Id          { get; set; }
+        public int      Id          { get; set; }
         public string   Nombre      { get; set; }
         public string   Apellido    { get; set; }
-        public Int64   DNI         { get; set; }
-        public Int16      IdPerfil    { get; set; }
+        public int   DNI         { get; set; }
+        public int      IdPerfil    { get; set; }
         public string email { get; set; }
         public string pass { get; set; }
